django-ui/requirements.txt
blinker==1.3
certifi==2018.4.16
chardet==3.0.4
cloud-init==18.2
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
defusedxml==0.5.0
Django==2.0.4
django-allauth==0.35.0
idna==2.6
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
MarkupSafe==0.23
oauthlib==2.0.7
prettytable==0.7.2
pyasn1==0.1.9
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.1
python-debian==0.1.27
python-systemd==231
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.11
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
ssh-import-id==5.5
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
